+++ esp-hal/src/clock/clocks_ll/esp32s2.rs
@@ -69,7 +69,7 @@ pub(super) fn enable_wifi(enable: bool) {
     });
 }
 
-pub(super) fn reset_mac() {
+pub(super) fn reset_wifi_mac() {
     SYSCON::regs()
         .wifi_rst_en()
         .modify(|_, w| w.mac_rst().set_bit());
@@ -98,12 +98,4 @@ pub(super) fn init_clocks() {
     SYSCON::regs()
         .wifi_clk_en()
         .modify(|r, w| unsafe { w.bits(r.bits() & !WIFI_BT_SDIO_CLK | DPORT_WIFI_CLK_WIFI_EN) });
-}
-
-pub(super) fn ble_rtc_clk_init() {
-    // nothing for this target
-}
-
-pub(super) fn reset_rpa() {
-    // nothing for this target
 }