+++ esp-ieee802154/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `esp_ieee802154::Ieee802154::new` no longer requires the `RADIO_CLK` peripheral (#3687)
+
 - MSRV is now 1.88.0 (#3742)
 
 ### Fixed